Made some syntax improvements, and added in basic snippets
When there was a function in a cfif the coloring would go all wonky, so I tried to tighten that code up a bit.  Also tried to get the cffunction colouring on the start tag to match the other coldfusion tags.

(Setting the theme to Brilliance Black points out rendering errors very well)